backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
gunicorn>=23.0.0
python-multipart>=0.0.12

# Database
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
cloud-sql-python-connector[asyncpg]>=1.12.0

# Validation & Settings
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.1.0

# Auth & Security
python-jose[cryptography]>=3.3.0
passlib[argon2]>=1.7.4
cryptography>=44.0.0
pyotp>=2.9.0

# Storage
boto3>=1.35.0
google-cloud-storage>=2.18.0

# Cache
redis>=5.2.0

# Task Queue
celery>=5.4.0

# Logging
structlog>=24.4.0

# PDF Export
fpdf2>=2.8.0

# Rate Limiting
slowapi>=0.1.9

# HTTP Client
httpx>=0.28.0

# Testing (dev)
pytest>=8.3.0
pytest-asyncio>=0.24.0
aiosqlite>=0.20.0
httpx>=0.28.0

# Linting (dev)
ruff>=0.8.0
mypy>=1.13.0
